+++ src/main/java/com/banchan/model/dto/questions/QuestionReportRequestDto.java
@@ -1,0 +1,20 @@
+package com.banchan.model.dto.questions;
+
+import com.banchan.model.entity.Reports;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class QuestionReportRequestDto {
+
+    private Integer userId;
+    private Integer questionId;
+
+    public Reports toQuestionReportEntity() {
+        return Reports.builder()
+                .userId(userId)
+                .questionId(questionId)
+                .build();
+    }
+}
